@@ -1,12 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 32
 
     defaultConfig {
         applicationId "com.bigkoo.pickerviewdemo"
-        minSdkVersion 14
-        targetSdkVersion 26
+        minSdkVersion 21
+        targetSdkVersion 32
         versionCode 1
         versionName "1.0"
     }
@@ -16,9 +16,10 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-    lintOptions{
-        checkReleaseBuilds false
+    namespace 'com.bigkoo.pickerviewdemo'
+    lint {
         abortOnError false
+        checkReleaseBuilds false
     }
 }
 
@@ -27,6 +28,7 @@ dependencies {
     implementation project(':pickerview')
 //   implementation 'com.contrarywind:Android-PickerView:4.1.9'
 //    implementation 'com.contrarywind:wheelview:4.1.0'
-    implementation 'com.android.support:appcompat-v7:26.1.0'
+    implementation 'androidx.annotation:annotation:1.3.0'
     implementation 'com.google.code.gson:gson:2.7'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
 }
