@@ -1,17 +1,15 @@
 apply plugin: 'com.android.library'
-apply plugin: 'com.github.dcendents.android-maven'
+//apply plugin: 'com.github.dcendents.android-maven'
 //apply plugin: 'com.novoda.bintray-release'//添加JCenter插件
 
 
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 32
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 26
-        versionCode 34
-        versionName "4.1.9"
+        minSdkVersion 21
+        targetSdkVersion 32
     }
     buildTypes {
         release {
@@ -19,7 +17,8 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-    lintOptions {
+    namespace 'com.bigkoo.pickerview'
+    lint {
         abortOnError false
     }
 }
@@ -46,5 +45,6 @@ dependencies {
 //    compile fileTree(include: ['*.jar'], dir: 'libs')
     api project(path: ':wheelview')
 //    api 'com.contrarywind:wheelview:4.1.0'
-    implementation 'com.android.support:support-annotations:28.0.0'
+    implementation 'androidx.annotation:annotation:1.3.0'
+//    implementation 'com.github.candRabbit:Android-PickerView:1.0.0'
 }
