@@ -1,7 +1,7 @@
 package com.bigkoo.pickerview.builder;
 
 import android.content.Context;
-import android.support.annotation.ColorInt;
+import androidx.annotation.ColorInt;
 import android.view.View;
 import android.view.ViewGroup;
 
